change stream to future

+++ lib/data/repository/supabase_profile_repository.dart
@@ -79,6 +79,7 @@ final class SupabaseProfileRepository implements ProfileRepository {
         .from('follows')
         .delete()
         .eq('followed_profile_id', profileId)
+        .eq('following_profile_id', supabaseClient.auth.currentUser!.id)
         .then((value) {
       Logger().d('正常にフォローを解除しました。');
     }).catchError((error) {
@@ -88,16 +89,20 @@ final class SupabaseProfileRepository implements ProfileRepository {
   }
 
   @override
-  Stream<bool> isFollowing(String profileId) {
+  Future<bool> isFollowing(String profileId) {
     Logger().d('isFollowing.profileId: $profileId');
     return supabaseClient
         .from('follows')
-        .stream(primaryKey: ['id'])
+        .select('followed_profile_id')
         .eq('followed_profile_id', profileId)
-        .map((event) => event.isNotEmpty)
-        .handleError((error, stackTrace) {
-          Logger().e('isFollowing.error: $error stackTrace: $stackTrace');
-          throw error;
-        });
+        .eq('following_profile_id', supabaseClient.auth.currentUser!.id)
+        .limit(1)
+        .then((value) {
+      Logger().d('isFollowing.value: $value');
+      return value != null && value.isNotEmpty;
+    }).catchError((error) {
+      Logger().e('isFollowing.error: $error');
+      throw error;
+    });
   }
 }
